+++ src/cmds/thinking.js
@@ -1,12 +1,8 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { EmbedBuilder, WebhookClient } = require('discord.js');
+const { EmbedBuilder } = require('discord.js');
 const { thinking } = require('../config.json');
-const config = require('../../config');
 
-const cmdHook = new WebhookClient({
-	id: config.cmdHook.id,
-	token: config.cmdHook.token
-});
+const { cmdHook } = require('../utils/WebhookManager');
 
 const thinkingData = new SlashCommandBuilder()
 	.setName("thinking")
